+++ src/Ethereum/ABI/Parameters.cpp
@@ -35,7 +35,7 @@ void ParamSet::addParams(const std::vector<std::shared_ptr<ParamBase>>& params) 
 }
 
 bool ParamSet::getParam(int paramIndex, std::shared_ptr<ParamBase>& param_out) const {
-    if (paramIndex >= _params.size() || paramIndex < 0) {
+    if (static_cast<std::size_t>(paramIndex) >= _params.size() || paramIndex < 0) {
         return false;
     }
     param_out = _params[paramIndex];
@@ -47,7 +47,7 @@ std::shared_ptr<ParamBase> ParamSet::getParamUnsafe(int paramIndex) const {
         // zero parameter, nothing to return.  This may cause trouble (segfault)
         return nullptr;
     }
-    if (paramIndex >= _params.size() || paramIndex < 0) {
+    if (static_cast<std::size_t>(paramIndex) >= _params.size() || paramIndex < 0) {
         // invalid index, return the first instead of nullptr
         return _params[0];
     }
